move TaskView and JobsView into controller package
